backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
pillow==10.4.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
httpx==0.27.2
firebase-admin==6.5.0
